fix(gym_futbol): check_env() warning about observation space - use float64 instead of float32

+++ gym_futbol/envs_v1/futbol_env.py
@@ -91,10 +91,10 @@ class Futbol(gym.Env):
         # [3] y velocity
         self.observation_space = spaces.Box(
             low=np.array([-1., -1., -1., -1.] *
-                         (1+self.number_of_player * 2), dtype=np.float32),
+                         (1+self.number_of_player * 2), dtype=np.float64),
             high=np.array([1., 1., 1., 1.] *
-                          (1+self.number_of_player * 2), dtype=np.float32),
-            dtype=np.float32)
+                          (1+self.number_of_player * 2), dtype=np.float64),
+            dtype=np.float64)
 
         # create space
         self.space = pymunk.Space()
